调试PID

+++ User/motor.h
@@ -1,10 +1,10 @@
 #ifndef __MOTOR_H
 #define __MOTOR_H
 
-extern float vP,vI,vD;
+extern int vP,vI,vD;
 void Motor_Init(void);
 void Motor_Speed(int speed);
-void Motor_velocity_control(float current_speed,float want_speed);
+void Motor_velocity_control(int current_speed,int want_speed);
  
 #endif
